+++ server/src/ClientToDBAPI/LogsRoute/mainRoute.ts
@@ -3,7 +3,7 @@ import { Router } from 'express';
 import HandleDeveloperRequest from '../../middlewares/HandleDeveloperRequest';
 
 import Routes from './routes';
-import logController from './Log';
+import logController from './log';
 
 const router = Router();
 router.use(HandleDeveloperRequest);
